Move PrivateRoute and PublicRoute to separeted files
