+++ w-books/src/config/routes.js
@@ -1,43 +1,15 @@
 import React from 'react';
-import {
-  BrowserRouter as Router,
-  Route,
-  Redirect
-} from 'react-router-dom'
+import { BrowserRouter as Router } from 'react-router-dom'
 import Home from '../App/screens/Home/index.js'
 import BookDetail from '../App/screens/BookDetail/index.js';
 import Login from '../App/screens/Login/index.js';
 import Header from '../App/components/Header/index.js';
+import PublicRoute from './components/PublicRoute/index.js'
+import PrivateRoute from './components/PrivateRoute/index.js'
 
 const isAuthenticated = () => {
   return localStorage.getItem('access_token') !== null;
 }
-
-const PrivateRoute = ({ component: Component, ...rest }) => (
-  <Route {...rest} render={props => (
-    isAuthenticated() ? (
-      <Component {...props}/>
-    ) : (
-      <Redirect to={{
-        pathname: '/login',
-        state: { from: props.location }
-      }}/>
-    )
-  )}/>
-)
-
-const PublicRoute = ({ component: Component, ...rest }) => (
-  <Route {...rest} render={props => (
-    !isAuthenticated() ? (
-      <Component {...props}/>
-    ) : (
-      <Redirect to={{
-        pathname: '/dashboard',
-        state: { from: props.location }
-      }}/>
-    )
-  )}/>
-)
 
 function AppRoutes() {
   return (
